requirements.txt
asgi-redis==1.1.0
asgiref==3.5.0
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
black==22.1.0
cffi==1.15.0
channels==3.0.4
click==8.0.3
colorama==0.4.4
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
Deprecated==1.2.13
Django==4.0.2
django-appconf==1.0.5
django-ckeditor==6.2.0
django-compressor==3.1
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-filter==21.1
django-js-asset==2.0.0
django-libsass==0.9
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
libsass==0.21.0
Markdown==3.3.6
msgpack-python==0.5.6
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
pytz==2021.3
rcssmin==1.1.0
redis==4.1.3
rjsmin==1.2.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
timedelta==2020.12.3
tomli==2.0.1
Twisted==22.1.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing_extensions==4.1.0
tzdata==2021.5
wrapt==1.13.3
zope.interface==5.4.0
